third_party/blink/web_tests/virtual/first-party-sets/http/tests/inspector-protocol/issues/third-party-cookie-blocking-first-party-set-expected.txt

Test third-party cookie blocking between sites in same First-Party Set
Test started

Running test: thirdPartyCookieBlockingDisabled
Number of cookies: 0
Number of cookies: 1
Cookie not blocked: true

Running test: thirdPartyCookieBlockingEnabled
Number of cookies: 0
Number of cookies: 1
Cookie blocked: true
Reasons: UserPreferences,ThirdPartyBlockedInFirstPartySet
{
    issue : {
        code : CookieIssue
        details : {
            cookieIssueDetails : {
                cookie : {
                    domain : cookie.test
                    name : foo
                    path : /
                }
                cookieExclusionReasons : [
                    [0] : ExcludeThirdPartyCookieBlockedInFirstPartySet
                ]
                cookieUrl : https://cookie.test:8443/inspector-protocol/resources/iframe-third-party-cookie-child.php
                cookieWarningReasons : [
                ]
                operation : ReadCookie
                request : {
                    requestId : <string>
                    url : https://cookie.test:8443/inspector-protocol/resources/iframe-third-party-cookie-child.php
                }
                siteForCookies : https://firstparty.test/
            }
        }
    }
}
